[IMP] helpdesk: add quick assignment in ticket kanban view
Before this commit, the user cannot assign someone in the ticket in the
kanban view as we can in the kanban view of project and task for
instance.

This commit displays the user_id field in the kanban view of ticket even
if the no user is assigned. By doing that, the user can assign someone
in the ticket without having to open the form view or another view.

task-4793688

closes odoo/enterprise#85510

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>