[FIX] project_mrp_workcenter_account: udpate analytic account with correct data from MO operation
**Problem:**
lines of mrp.workcenter.productivity added from
the workorder widget do not update the linked
analytic account with the right amount

**Steps to reproduce:**
- enable "analytic accounting" setting
- Unarchive MTO route
- Create a new product
- In the inventory page check MTO and Manufacture routes
- click on the Bill of materials smart button and create
a new BOM
- add a component and an operation
- Create another product, set "product type" as service
- in the "create on order" field select "project & task"
- create a new quotation with these two products and confirm
- click on the "manufacturing" smart button
- in the work orders page click on the "open work order"
button at the right of the line
- click on add a line
- set a duration of 30:10 and clik elsewhere on the screen
- save
- click on the "analytic account" smart button
- select the only line and click on the "gross margin"
smart button

**Current behavior:**
There is a single line with an amount of 0.01

**Expected behavior:**
The price should reflect the hourly rate of the employee
for the task (that's 50.28 with a 100$/hours rate for
instance)

**Cause of the issue:**
There is two issue here.
First Issue :
When clicking on save on the widget, this will
modify the time_ids field of the mrp.workorder
this will create a new mrp.workcenter.productivity
with the values we wrote.
Consequently, on the vals_list when the create method
is triggered duration is set to the duration we wrote
(so here 30.17)

Because of this, when the _compute_duration is the triggered
previous_durations will be [30.17] and _create_analytic_entry
will be called with a parameter of 30.17
https://github.com/odoo/enterprise/blob/8a21b4f8ee8d24ebbc9ac2eb08696a0de056357f/project_mrp_workorder_account/models/mrp_workcenter_productivity.py#L10-L14

Therefore, inside _create_analytic_entry
duration will be 0 and amount will also be 0

This issue does not happen if the duration is set by changing
the "real duration" in the "work orders" page of the manufacturing
order.
This is because when doing it this way, we modify the duration
of the mrp.workorder which triggers the set_duration_method
which also creates a new mrp.workcenter.productivity but
here the _prepare_timeline_vals returns a dictonary without
a duration field
https://github.com/odoo/odoo/blob/4fe90167266a92c3a5941eac9a6a2084a80056ac/addons/mrp/models/mrp_workorder.py#L352-L353

Second Issue:
The line we see with a 0.01 amount is actually created
before we save.
It's created from the onchange method when we change the
duration on the form.
The onchange of duration triggers an onchange of start date
which gets the value of duration and triggers of recomputation
of duration
https://github.com/odoo/odoo/blob/4fe90167266a92c3a5941eac9a6a2084a80056ac/addons/mrp/models/mrp_workcenter.py#L522-L531
As a consequence, _compute_duration is called.
self.duration is initially not rounded so in our exemple
previous duration will be equal to 30.1666666666668
https://github.com/odoo/enterprise/blob/8a21b4f8ee8d24ebbc9ac2eb08696a0de056357f/project_mrp_workorder_account/models/mrp_workcenter_productivity.py#L10-L11
However after the call to super()._compute_duration(), self.duration
will be rounded thanks to the use of convert to duration inside
the super method
https://github.com/odoo/odoo/blob/4fe90167266a92c3a5941eac9a6a2084a80056ac/addons/mrp/models/mrp_workcenter.py#L472

So when doing the difference here inside _create_analytic_entry
https://github.com/odoo/enterprise/blob/8a21b4f8ee8d24ebbc9ac2eb08696a0de056357f/project_mrp_workorder_account/models/mrp_workcenter_productivity.py#L38
duration will be 5.555555555556424e-05 and amount will be
-0.005555555555556424 which results in the creation of this line
with an amount of 0.01

**fix:**
Calling the creation of account.analytic.line from the compute method leads to other problems :
For instance, if a line is added in the "open work order" widget and we change the end date but change our mind and don't save the line, the duration will have been computed and a new account analytic line will have been created

Whereas calling the method _create_analytic_entries from create() and write() ensures it's only called when the model is saved

opw-4715690

closes odoo/enterprise#88117

X-original-commit: e4ed63dd83783e3e430b3ccc693193e19c6714d3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Pierre-Louis Hance (plha) <plha@odoo.com>