[FIX] knowledge: assign correct res_id to cover image attachment
The overhaul of html editor into knowledge introduced an issue
involving the auto-vaccuum. The following commit passes the
res_id value of the currently existing knowledge.cover record:
https://github.com/odoo/enterprise/commit/08d84f8b61450a0ebd0d2aadfe8f227bda283edd
However, when we set a new cover image, we will always create a new cover
record and the new cover image attachments are linked to the previous
cover record.

The auto-vaccuum then deletes unused cover images and any attachments linked,
which will delete the actual cover image.
The web_unsplash/attachment/add does not need to pass in a res_id of the
cover image since we always create a new record at the end of the workflow.

opw-4629300

closes odoo/enterprise#85785

X-original-commit: ae0f7797888775bf1882c26b9932dfac58697c40
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Piero Calenzani (pca) <pca@odoo.com>