[IMP] documents: add a new pictogram and use it in portal
This commit introduces a new pictogram to the existing
ones, in order to replace the Documents app icon currently in use within
Portal.

Prior to this commit, Documents was using its module icon on the Portal
homepage, which was inconsistent with the other icons displayed on the
page, which were all using pictograms from our library.

task-4793934

closes odoo/enterprise#85611

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>