[IMP] iot: show escpos errors on test print
In the community PR (odoo/odoo#208611) we have added device events for
errors and warnings from ESCPOS receipt printers (e.g. low on paper).
This commit adds support for displaying these errors to the 'Test'
button on the device form.

NB:
At the moment all the logic is contained directly in the device form
class, however once we add this functionality to the POS it would be
good to have the error logic in a common place, and re-use it here and
in the POS.

task-4756311

closes odoo/enterprise#85573

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>