[IMP] account_reports: Show VAT report pdf from the entry
There are 3 improvements here:
- Showing the attachments of the return on the journal items of
  the return
- Showing the Accounting date by default on draft entries check
- Preventing the Opening Date field to be red when creating a
  Accounting Periods wizard

In version 18.2, when a closing entry move was generated, the PDF (and
other files such as XML) would be displayed on the journal items view.
In version 18.3, the Tax returns were improved but the attachment was
not displayed on the journal items line of the entry of the return
anymore.

The second improvements was to display the Accounting Date on the
Draft entries that needs to be reviewed in a Tax return line to give
the user more context on why they need to review those specific Draft
entries.

Lastly, when accessing Tax Returns it would ask for us to set the
Accouting Periods through a wizard which prefilled informations such
as the VAT Periodicity but the Opening Date needs to be filled by this
user but this field would show as red when opening the wizard as it is
required and no information was prevent. To allow the field to be red
only if we try to submit the informations we are passing the
informations to the wizard as default value which wouldn't trigger the
wizard to try to save values when being created.

task-4784283

closes odoo/enterprise#85570

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>