[FIX] ai: fix error when closing agent chat
Before when closing agent chat an exception occurred because the
`channel_id` was undefined. This happens because when doing an await in
the thread is set to `undefined`.

This commit introduces a few modifications to the behaviour of the chat
with the agent:

Firstly, this commit moves the code calling the ORM to delete the chat
history when closing the chat window to the
`chat_window_model_patch.js`. This allows for the behaviour to also be
applied when closing the chat from the chat bubble. Additionally, the
function has been modified to ensure that the thread is defined when the
ORM call is made.

The ORM call to close the agent chat was simplified to use the already
existing `close_ai_chat` from the `discuss.channel` override. The method
now uses a new `channel_type` "ai_chat", to identify channels which
contain an `ai.agent`.

Along with the new `channel_type`, a new function `_get_or_create_ai_chat`
was introduced to properly create the "ai_chat" channels.

task-4796898

closes odoo/enterprise#88794

X-original-commit: eb9451740a97a58383218230c48daddb843ede69
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: Dylan Proovost (DPRO) <dpro@odoo.com>