[REM] mrp_plm: remove default ECO type 'New Product Introduction'
There are two ECO types by default, which can be confusing, and PLM isn't
intended to introduce new products in the first place, so we get rid of
'New Product Introduction' records from the default data.

Note: This commit will not be accompanied with an update script to
remove the records, in order to preserve the previous behaviour in
existing databases.

Task ID: 4664077

closes odoo/enterprise#85596

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>