[FIX] account_iso20022: support pain.001.001.09 for ISO20022
pain.001.001.09 is a newer version of the ISO20022 format.
Before this commit, we only supported it for SEPA Credit Transfer,
using the same selection field as the one allowing to choose
national variants of SEPA. This fix only intends to solve the issue
for existing customers by using a system parameter that needs to be
switched from "False" to "True". A cleaner solution will come in the
future, after a refactoring of the payment methods.

task-4647016

closes odoo/enterprise#86235

X-original-commit: b194dc24e311336c432a9973212f071f94bc9b27
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>