[FIX] account_followup: select right template to send followups through snailmail
Steps to reproduce:
- Have an overdue invoice
- Go to Accounting > Customers > Follow-up Reports and click the partner
- Change the state of the report to something different than the first state
- Click "Follow-up", select "By post" and send
- Settings > technical > Email > Snailmail Letters find the letter corresponding to the report
- Check the PDF, it is using the mail template of the first followup level

Cause:
When sending a followup report, the followup level of the report is recomputed for it to have the correct value.
Then the report is generated.
But the option to send a letter uses methods from the module `snailmail` for example `` which don't have options. The options are then recomputed to generate the report, reading [`partner.followup_line_id`](https://github.com/odoo/enterprise/blob/c8dac822cf3df23922488f33c07a114bbea05198/account_followup/models/account_followup_report.py#L42) so getting the wrong value.

Solution:
Generate the report in `_send_snailmail`, then create the `snailmail.letter` object from the attachment. This way the options are not lost in the creation process.

Created a new method to return the attachment as `_print_followup_letter` also needs it.

opw-4766804

closes odoo/enterprise#87362

X-original-commit: 8ff3670f38208b9bc4ff020c82a182a286713d2c
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>