[FIX] appointment: resolve attendee invitation error
In version 18.0, a traceback occurs when editing the 'Appointment:
Attendee Invitation' email template and then signing up for an
appointment.

**Steps to reproduce:**
- Go to Appointments.
- Click the "Action" gear icon next to an appointment, then select Edit.
- Navigate to the Options tab.
- Hover over the Booking Email template and click the arrow icon to open
  it.
- In the Content tab, add some text near "Hello", then click Save.
- Return to the Appointment form
- Click the "Go to Website" smart button at the top
- Select a date and time.
- Confirm the appointment.

This raises the following traceback:

```
SyntaxError: t-elif directive must be preceded by t-if or t-elif directive
Template: <Element div at 0x775b92a07840>
Path: /div/div/t[8]
Node: <t t-elif="target_responsible"/>
```

In saas-18.3, editing the 'Appointment: Attendee Invitation' email
template throws an error when saving, making the template uneditable.

**Issue:**

The issue originates from this [commit], which introduced a
<div> inside a <p>—a structure that violates the HTML specification.
Editing anything near the word 'Hello' in the template causes the
<p> tag to close prematurely during rendering. This pushes the <div>
outside of the <p>, breaking the directive chain, which results in
the traceback error.

Previously, the <div> was also pushed outside the <p>, but only after
evaluating all t-if, t-elif, and t-else conditions. Hence, it worked
correctly without editing the template.

**Solution:**

Placed the <span> as a child of the <p> tag to resolve the issue.

This fix also resolves the issue of being unable to edit the template
"Appointment: Attendee Invitation" in the UI on saas-18.3.

[commit]: https://github.com/odoo/enterprise/commit/8b9a1a4462453bf64e7bc1307ededd0d127a27fa

opw-4624645

closes odoo/enterprise#86965

X-original-commit: f52cb69324b0ad240d723343ded741e36ca77773
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Hiral Bhavsar (hbh) <hbh@odoo.com>