[FIX] helpdesk: auto-subscribe recipients to tickets on share and message
Before this commit, when a ticket was shared through the share wizard or when a
message was sent to a customer via the chatter, the recipients were not added
as followers of the ticket. As a result, they couldn't see the ticket in the
portal.

This commit ensures:
 - Recipients are now added as followers of the ticket when it is shared with
   them through the wizard.
 - A message to a customer in the chatter also adds them as a follower of the
   ticket.

task-4781259

closes odoo/enterprise#91019

X-original-commit: 49c89a7f2e1171d9e683102ac207115ed3479c1c
Related: odoo/odoo#220552
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>
Signed-off-by: Maan Pankajbhai Patel (maap) <maap@odoo.com>