[FIX] *: adapt tests to new constraint on tax group accounts
In the related community commit, we added a constraint that requires the
Tax Receivable and Tax Payable accounts on tax groups to be a Receivable
or Payable account, be reconcilable and set to Non Trade.

In this commit we adapt the current tests that use accounts on tax
groups to have their accounts comply to the new constraint.

task-3763030

closes odoo/enterprise#87437

X-original-commit: 6b9ffe4c3f29067a69707349c8622818e124513a
Related: odoo/odoo#213823
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>