[FIX] test_marketing_automation, marketing_automation_whatsapp: prevent access errors when sending queued messages
When the `marketing_automation_whatsapp` module is installed, if a user
**without** the "Marketing Automation / User" group but **with** the
"WhatsApp / Administrator" group sends a WhatsApp message, the
"WhatsApp: Send In Queue Messages" cron may crash with an access error.
This happens because the system tries to access marketing traces with
the user that sent the message. However, marketing traces are restricted
to Marketing Automation users.

As a result, all WhatsApp messages in the same batch are blocked
indefinitely, even those that could otherwise be sent successfully.

opw-4678339

closes odoo/enterprise#85783

X-original-commit: 1b9b7eee5a969b64f981f934614bdd03dfc5d818
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>