[FIX] spreadsheet_sale_management: duplicate quote spreadsheets
1. Create a quote using a quote template that uses the quote calculator
   (e.g. the demo one, works fine),
2. Modify it using the quote calculator,
3. Duplicate it,
4. Open the quote calculator in the new quote
-> it's filled with default values instead of the one from the other quote
you dpulicated one could expect after duplication (and so if you save
without changing anything in the spreadsheet, you 'reset' the quote even
though it's probably not what you wanted to do),

Expected behaviour: the quote calculator spreadsheet is duplicated for the
new quote with the same values.

Implementation note:
Why not just `copy=True` on field `spreadsheet_ids`?

Because the spreadsheet should be copied, but not only!
The `spreadsheet.revision` linked to the spreadsheet also need to be copied.

If they are not copied, revisions which are not part of the last snapshot
will be lost on the new spreadsheet record and it won't lead to the
same spreadsheet data.

This is handle by an override of `copy`. Unfortunatly, it can be an
override of `copy_data` because revisions have a parent/child relationship
which forces to create them one at a time.

closes odoo/enterprise#86158

Task: 4801418
X-original-commit: 8e0099c3c510852264fa479163ff12ee92a1a10c
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>