[IMP] appointment,*: add is_default and 'phone' type to questions
*: appointment_crm,appointment_hr_recruitment,appointment_sms,
website_appointment

PURPOSE
=======

Allow users to customize their appointments more, by making
the phone question an appointment.question record instead of
a hard-coded question in the form template.

Also allow users to enable the new is_default boolean field on
the question, meaning that it will be included in the questions
of new appointment types on creation.

AFTER THIS COMMIT
=================

Phone is now a new question_type. It is not hard coded in the
attendee form in the front end as is email and name. Also, a
data record is now containing the default phone question, that
will be a default question on install. It will be included in
all appointments until deleted / modified. Questions can all be
configured as default questions with the new is_default field.

CHANGES
=======

A new is_default boolean field is added on the appointment.question
model. If no value is passed at creation, all default questions will
be added to a new appointment type (in the default order).

If at least a phone question is used, then the first phone question
in order will be used to prefill partner data, to match the partner
in the controller, and to update the partner if appropriate. The
prefill data based on the data of last booking is adapted in the JS.
The phone question is used to prefill data / update partner's phone,
see '_get_main_phone_question'.

The phone will now appear in the part of the description that includes
question answers. (It is kept as is, and will not be updated if the
partner's updated, unlike the name and email)

A new sql constraint ensures that default questions are always reusable
since it would not make sense otherwise. is_reusable is also turned into
a computed field in order to make it clearer / easier for the end user.

TESTS
=====

Tests are updated as well, mostly because 'phone' argument has now
disappeared from the controller argument list, since it is now treated
as other answers.

Two tests are added to ensure only the first phone question propagates
the phone to the partner, when appropriate. Also ensure that the
submission still work as expected when no phone question exists.

UPGRADE
=======

We add the default phone question as data, make it first in sequence,
and add it to all existing appointments (so that nothing changes on
migration)

UPG PR: https://github.com/odoo/upgrade/pull/7691

Task-4655040

Part-of: odoo/enterprise#82973
Related: odoo/upgrade#7691
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>