[IMP] account_online_{payment, synchronization}: improve the KYC flow
Users currently initiate the KYC
flow automatically when making a first payment.

This improves the KYC flow by prompting users to
complete it as soon as they add a bank account in
Odoo. We send them an email, create a notification
and schedule an activity to ensure they complete
the KYC process.

Task ID: 4487895

closes odoo/enterprise#85772

Related: odoo/upgrade#8141
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>