[IMP] l10n_co_{edi,dian}: support ICL taxes for DIAN edi
Like the sugar tax, alcohol is also taxed differently in Colombia such
that we need a special case for how to compute it. The Per Unit Amount
is the total tax amount divided by the alcohol percentage.

To facilitate this change, the field that was previously used for the
volume in ml for sugary drinks has been made into a generic "Specific
Component Nominal Tax" field which can be used for any extra tax data
depending on what type of tax is on the product.

task-4634313

closes odoo/enterprise#85930

X-original-commit: 4951e3bdaab824fd354d771b8ca15d2bd581221a
Related: odoo/odoo#210715
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Andrew Gavgavian (andg) <andg@odoo.com>