[FIX] account_accountant: apply Bank St line regex on narration
We don't want the user to differentiate the places where to apply the
regex search (reco model) on statement lines. We thus apply it on the
label, the transaction details and the narration altogether, returning
on the first match. We don't need the "from_transaction_details"
selection option anymore.

task-4749338

closes odoo/enterprise#85861

Related: odoo/odoo#210588
Related: odoo/upgrade#7701
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>