[FIX] stock_barcode_picking_batch: avoid useless RPC
Issue
=====
In `StockBarcodeKanbanRenderer` `onWillStart`, the python method
`get_model_records_count` is called to fetch the number of ready picking
or batch picking. The issue is this method is useless for manufacturing
order but is called anywway.
With this commit, this method is called only when needed.

How to reproduce
================
- Install Inventory, Manufacturing and Barcode;
- Inventory > Configuration > Settings > check "Batch, Wave & Cluster
  Transfers";
- Go in Barcode > Operations > Manufacturing: If you inspect calls,
  you'll see `get_model_records_count` is called.

closes odoo/enterprise#86366

X-original-commit: 59c6efb43d888e2cf59d19db300f5e7c42fa45ef
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>