[FIX] account_accountant: better handling of reco model regex errors
Before, with some errors in the regex search in the different statement
line attributes, the search would just return 0, when the regex was just
badly written, without giving any hint to the user. Now the method
will raise an user error telling the user that the regex is badly
written instead of returning 0.0 or a traceback (AttributeError was not
handled).

Examples of previously failing regexes:
- "([^\d]*)" -> AttributeError
- [\d,]* -> IndexError (was already handled by https://github.com/odoo/enterprise/pull/84284)
- ValueError shouldn't arise anymore with that same PR, so it is removed

closes odoo/enterprise#88246

X-original-commit: dd3c4583e54fa4a61eda1c5dbc78e9fb57258056
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>