[IMP] web_grid: make header of columns more locale friendly
The headers of the columns in the Grid view use a fixed format, which is
too restrictive for languages where the order of words changes, or where
additional literals like commas may or may not be present.

With this commit, we use the Intl API to ensure that we produce a valid result in all languages.
For instance, French becomes:

lun.
19 mai

instead of:

lun.,
mai 19

Task-4801503

Part-of: odoo/enterprise#85906
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>