[FIX] voip: no crash on activity rescheduling call activities (2)
Before this commit, when re-scheduling a call activity to a date
of today or older, it raised the following error:

```
Caused by: TypeError: Cannot read properties of undefined (reading 'avatarUrl')
at ActivityListPopoverItem.template
```

This happens because in this specific case of change of call activity
deadline date, the VOIP softphone code is fetching activity data of
today faster than odoo views, thanks to relying on bus notification,
whereas odoo views and chatter rely on mostly on returned RPCs.
The code of VOIP returns activity data with a custom formatter,
which omits `persona` that is important for the good templating
of an activity in Chatter otherwise there's the crash above.

A recent PR [1] attempted to fix this issue, but the syntax for
Store was wrong: it used the syntax for an item in field
list (e.g. in `_to_store_defaults`) instead of pure store data
from a specific record. This bad use of Store.One() resulted in
actually returning `False`.

[1]: https://github.com/odoo/enterprise/pull/84970

opw-4586756

closes odoo/enterprise#86122

X-original-commit: 53548694f8db1ed1095bf62a8a1aef90bbd3e2a6
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>