[FIX] industry_fsm: fixed runbot tour fail
The tours fails because when we edit the value in the input field of customer it
clicks on the field and then fills out the value and then a change event is
propagated and text based autocomplete results are shown (and m2o fields with
partner autocomplete wizard are a tad bit slower), after that tours moves to
next step and the new result dropdown menu is not loaded and it gets its
reference from older menu which got closed in between and tries click it and no
change is reflected but the new menu is not closed and tour continues to
industry_fsm_report where we create a worksheet from a many2one field using
create and edit so we click on the m2o input and wait for options but as partner
field loaded dropdown is loaded it clicks on click and edit trying to create a
new partner and opening thus fails the tour.

Fix:

- After entering the text to instantly load the dropdown by clicking on it and
give time for loading of menu.

task-4800194

closes odoo/enterprise#85985

X-original-commit: cbfed6c949f928e29d1980a0d64394adf045cc28
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: K Theja (kthe) <kthe@odoo.com>