[FIX] planning: fix typo in gantt renderer of planning
Before this commit, the code checks if `this.connectorDragState.draging`
if truly before using the split tool because we don't want to execute
that if the user is currently dragging a planning slot. The problem is
`draging` attribute does not exist in `connectorDragState`, it should be
`dragging` instead.

This commit fixes the typo to correctly check if the slot is dragging
before doing the split tool or not.

closes odoo/enterprise#86193

X-original-commit: ca92cdacb0d89298adc934acc7d1715a51df4cd2
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>