[IMP] web_gantt,planning: condense time displayed in pill display name
Before this commit, the time displayed to show start and end time of a
pill could be long in some languages, like in English when we use AM and
PM, because of that, there are not enough spaces to display other useful
information in the display name of the pill.

This commit changes the time format to display the time in HH:MM in the
display name of pill.

task-4801407

Part-of: odoo/enterprise#85945
Related: odoo/odoo#210758
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>