[FIX] planning: display allocated hours in gantt pill only if percentage != 100
Before this commit, when a shift has an allocated hours set, the
allocated hours is displayed in the gantt pill even if the allocated
percentage is at 100%.

This commit adds the allocated_percentage field in the gantt view to
make sure the allocated percentage is fetched to be able to know if the
slot has an allocated percentage set to 100% if yes then the allocated
hours will not be displayed in the display name of the gantt pill.
This commit also reviews a bit the code to display or not the allocated
hours to only call the parent method when it is really needed.

task-4801407

closes odoo/enterprise#86012

X-original-commit: 730343313726c2bbbed843f4a544cac2996c85a4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>