[FIX] l10n_mx_edi_pos: display CFDI usage on POS orders in backend
**Issue**
In the Mexican localization, the CFDI usage (`l10n_mx_edi_usage`) field is not visible on POS orders in the backend. This prevents users from selecting or modifying the usage when manually generating an invoice from the POS order.

**Steps to Reproduce**
1. Install the Point of Sale app and the Mexican localization (l10n_mx)
2. Complete a POS order without generating an invoice
3. In the backend, go to Point of Sale > Orders > Orders
4. Open the newly created order
5. Notice that the CFDI Usage field is missing

**Root Cause**
The usage field is given a default value at the model level:
https://github.com/odoo/enterprise/blob/bc3db24f83473d5646f7c2cfca8ed1c5b064ea2e/l10n_mx_edi_pos/models/pos_order.py#L65-L70
However, this value gets lost during the order processing flow. When an invoice is requested directly from the POS UI, the usage is explicitly set in the `sync_from_ui` method: https://github.com/odoo/enterprise/blob/bc3db24f83473d5646f7c2cfca8ed1c5b064ea2e/l10n_mx_edi_pos/models/pos_order.py#L110-L112
But if no invoice is requested at the time of the order, the usage is not preserved or restored in the backend.
**Fix**
Explicitly set the CFDI usage to its default value when an invoice is not requested during the POS flow. This ensures the usage field is retained and displayed properly in the backend, allowing users to create invoices manually.

opw-4776487

closes odoo/enterprise#86345

X-original-commit: 138cda3157d045c62a96c25adfbde2b3c43cc109
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Carmine Santomauro (csan) <csan@odoo.com>