[FIX] social_instagram: allow facebook crawler user-agent in robots.txt
Lately we have seen that when trying to upload to Instagram throught the
media link, the request is blocked by the robots.txt file. This is
because Facebook crawler is trying to access our image url and it is
not being able to get the image in their side.

This commit tries to add this user-agent to the robots.txt for both
when we have and do not have website installed, without altering any of
the existent templates. This will silently add it for website, so the
user will be able to modify the robots.txt file if needed, but won't be
able ot remove it.

Note: We only allow the facebook crawler user-agent where is striclty
neeeded, so we avoid having a noticeable imapact on the website
bandwidth. (sometimes facebook abuses where this agent is allowed, and
tries to crawl more than what's necessary)

opw-4726371

closes odoo/enterprise#88896

X-original-commit: faf9c834b6108d416c9d477ad29367176d16fe8c
Related: odoo/odoo#216508
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>