[FIX] payment_sepa_direct_debit: mandate validation with bank sync
Currently the SEPA mandate auto validation is done when the payment ref of the
statement line is the same as the name of the payment transaction and that the
partner id matches between SEPA mandate and statement line.

Issue is that some flows (at least the bank synchronization handled through
"account_online_synchronization" and odoofin) do not set the partner id on the
bank statement line (until reconciliation) but provide it first through the
"partner_name" field.

To ensure those flows work fine with SEPA, we should also match transactions and
statement lines whose partners names match when there is no "partner_id" set on
the bank statement line.

opw-4536189

closes odoo/enterprise#86495

X-original-commit: 61564370d4f661d9995867ead2367c30add2b589
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>