[FIX] l10n_nl_reports: update cron to fetch status
The way the cron was used before was by manually triggering it from
the Tax Report submission flow. The cron was set to not never trigger
otherwise (9999 months). However, if, for whatever reason, the cron
fails/crashes, the status of the submitted report would never be fetched
unless a report is submitted again.

closes odoo/enterprise#85986

Related: odoo/upgrade#7712
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>