[IMP] account_reports: show account name when account_code grouping
In the current consolidation, you can make multi-company accounts and
assign a code for each company.
But a code must be unique inside each company.  So it is not designed to
map multiple accounts from a second company into one account of a first
company.

It is actually possible to achieve that effect by not making the
accounts of the second company belong to the first company, and still
make the code mapping.
So you can make several accounts from the second company, map
to the same code for the first company.

Thus grouping by account codes become interesting for the accounting
reports, in order to get a view of that consolidation.

In order to make this reporting more clear, the name of the account
corresponding to the account code (in the current company) is now shown
when grouping by account code.

task-4801891

closes odoo/enterprise#93821

X-original-commit: a68e075c0b5695d10daaa32db9253c6ac00f2946
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Arnaud Sibille (arsi) <arsi@odoo.com>