[FIX] documents: fix backend requests upload
Before this commit, when one tries to upload a file to a document request for
the first time it does nothing. This is because we call an async memoize method
which causes the file to be cleared between the memoize rpc call and the file
processing after that call... After that first call the user can upload a new
file without issue but it's still annoying.
This commit fix this issue by storing the file in a variable before doing the
memoize rpc call.

Task-4778429

closes odoo/enterprise#88751

X-original-commit: f57850f382a132722b0481e3f437593bb0108539
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>