[IMP] industry_fsm,project_enterprise: add show sub-tasks button in control panel
Before this commit, the behavior was not really the one expected, the
sub-tasks should not be available in the tasks kanban views of a
project, those sub-tasks should be hidden by default and only visible if
the user wants to see them.

This commit adds a new button to let the user choose if he wants to see
or not all sub-tasks linked to his project in his pipeline. That button
will be integrated inside the main button uses to display the top menu.
That is, now the button to show/hide the top menu is converted into a
dropdown to display the button to show/hide sub-tasks and the one to
show/hide top menu. By doing that, the both buttons will be easily
discoverable and also easier understandable for the users.
This commit also removes the filter "Hide sub-tasks" since now the new
button will do the same job.
By default, the sub-tasks inside the same project of their parent task
will no longer be display in the task main views of a project, the user
will have to click on the new button to show the sub-tasks. That new
button will store in the user session to remember the user choice for
all projects.

task-4801249

closes odoo/enterprise#87066

Related: odoo/odoo#213096
Related: odoo/upgrade#7806
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>