[IMP] sale_subscription: define end_date on pdf quote
Before:
The subscription end date was not displayed on the quote PDF, requiring users to
manually adjust end of the subscription period.

After:
If a subscription end date is defined, it is now displayed next to the recurring
plan on the quote PDF, reducing manual effort and improving clarity for
customers.

task-4801402

closes odoo/enterprise#86161

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>