[FIX] l10n_cl_edi: remove trailing space dirorigen
Steps to Reproduce :

1. Install l10n_cl_edi.
2. Navigate to Contacts create an invoice for CL Company
3. Choose the document type: (33) Electronic Invoice.
4. Confirm the invoice.
5. Ensure the partner has only the street value set
   (e.g., "Name"), and street2 is empty.

Issue:

When generating the DTE, if street2 is empty, a trailing
space appears in the tag
ex: DirOrigen>Name </DirOrigen>

Solution:

Applied strip() to remove the extra space when street2 is empty.
This ensures no trailing space when street2 is empty and proper
spacing when both street and street2 is present.

opw-4765451

closes odoo/enterprise#86842

X-original-commit: ed0831965b2c88db1e0a25d65881aca9987e1bfd
Signed-off-by: Raju kumar Thakur (rkth) <rkth@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>