[FIX] hr_payroll: make the company_id as readonly in payslip
Currently a traceback is occurring when the user tries to create a Draft Entry,
in the batch that contains payslips of multiple companies.

To reproduce this issue:

1) Install hr_payroll_account and switch to multi-company environment
2) Create two employees with two running contracts respectively
3) Now create a Batch record from payroll/payslips/batches
4) Create two playslips with different companies with the created batch
5) Compute both the payslips and open the batch
6) Now, try to click on the  Create Draft Entry

Error:
A traceback will occur.
"ValueError: Expected singleton: res.company(1, 2)"

Cause:
When the user created multiple payslips with multiple companies containing
the same batch and tries to create a Draft Entry, we filtered out the payslips
based on the payslip state, and try to change the state to Done using the
action_payslip_done method.
https://github.com/odoo/enterprise/blob/1e5d7155466f80942f0b8fe2df7fdfb86cae7417/hr_payroll/models/hr_payslip_run.py#L70

So the self will contain multiple payslip recordsets with different company in
each payslip. This leads to the above traceback when accessing company_id from
self
https://github.com/odoo/enterprise/blob/1e5d7155466f80942f0b8fe2df7fdfb86cae7417/hr_payroll_account/models/hr_payslip.py#L61

Solution:

In payslips, the company_id field should be in readonly state, since the value
of the company_id will  computed from the employee_id which is a required field.

Also raised a validation Error for the previously created record to handle the
traceback.

opw-4797089

closes odoo/enterprise#89539

X-original-commit: 6b856732dc73a2d34af7c070432b00008124bfca
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>