[FIX] pos_self_order_iot: sign communication fails
With the introduction of signing communication between the IoT box
and the DB, a `sign_communication` method is called on the `iot.box`
model for every IoT request. This works fine when there is a logged in
user, however in the case when there is no user it fails. This is due
to the `call_kw` route used by ORM calls requiring a user.

In the Kiosk environment, there is no logged in user, so the request
fails. To fix this, a new route has been added to proxy the request,
which verifies the session ID and then returns the authorisation hash
as usual.

task-4809897

closes odoo/enterprise#86093

X-original-commit: 1c6ff0efca3f0d2a465a201c74df7e690150f398
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>