[FIX] knowledge: load `display_name` field for `MoveArticleDialog`
The `MoveArticleDialog` component uses the display name to indicate
which article will be moved. However, this field is not loaded on the
record object, resulting in an undefined value. As a result, the modal
incorrectly displays the article as untitled.

Steps to reproduce the issue:
1. Open an article in Knowledge
2. Make sure that the article has a title
3. Click on the "Move To" option from the kebab menu
=> The modal says "Move the untitled article under:"

TO BE: The modal should say: "Move <display_name> under:"

To fix the issue, we will add the `display_name` field to the field
dependencies of the `KnowledgeTopbar` component.

Task-4636494

closes odoo/enterprise#87211

X-original-commit: 2a2facc4ef16efc2a090147e0fe2de7b9f2f4607
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>